Create shared MapKeyView
